Fix up related to Rating. Add navbar Notification and Chat (on progress)

+++ back-end/src/main/java/com/graduate/thesis/backend/entity/UserProfile.java
@@ -70,5 +70,5 @@ public class UserProfile {
 
     private int ratingValue;
 
-    private float ratingAverage;
+    private double ratingAverage;
 }
